requirements.txt
pydantic>=2.6
beautifulsoup4>=4.12
langdetect>=1.0.9
PyYAML>=6.0
pandas>=2.0
numpy>=1.24
tqdm>=4.66
scikit-learn>=1.3
sentence-transformers>=2.6
lxml>=4.9
python-dateutil>=2.8
charset-normalizer>=3.3
pytest>=8.0
python-docx>=1.1.2
